Updated Nav and Footer
Gearing up for an 'upgrade' of the afya innovation lab features.
Added a new Nav bar with new routes.
Added a new footer with other new routes.
All aimed at ending up with a full stack dynamic site.
Intending for a registration page, a subscribe functionality using phone and or email, an events page that connects to a database, a health news page that could fetch from a purely health focused news API, etc., a resources page that hosts learning videos and some learning content and/or notes